refactor: JWT 토큰과 FCM 토큰의 분리
