+++ src/main/java/umc/stockoneqback/auth/controller/TokenReissueApiController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import umc.stockoneqback.auth.controller.dto.request.ReissueTokenRequest;
-import umc.stockoneqback.auth.domain.model.AuthToken;
-import umc.stockoneqback.auth.domain.model.TokenType;
+import umc.stockoneqback.auth.domain.model.jwt.AuthToken;
+import umc.stockoneqback.auth.domain.model.jwt.TokenType;
 import umc.stockoneqback.auth.service.jwt.TokenReissueService;
 import umc.stockoneqback.auth.utils.TokenResponseWriter;
 import umc.stockoneqback.global.annotation.ExtractToken;
@@ -21,8 +21,7 @@ public class TokenReissueApiController {
     private final TokenResponseWriter tokenResponseWriter;
 
     @PostMapping
-    public ResponseEntity<Void> reissueTokens(@ExtractToken(tokenType = TokenType.REFRESH) final String refreshToken, final HttpServletResponse response
-    ) {
+    public ResponseEntity<Void> reissueTokens(@ExtractToken(tokenType = TokenType.REFRESH) final String refreshToken, final HttpServletResponse response) {
         final AuthToken authToken = tokenReissueService.invoke(new ReissueTokenRequest(refreshToken));
         tokenResponseWriter.applyToken(response, authToken);
