--- src/main/java/umc/stockoneqback/auth/domain/model/TokenType.java
@@ -1,5 +1,0 @@
-package umc.stockoneqback.auth.domain.model;
-
-public enum TokenType {
-    ACCESS, REFRESH
-}
