+++ src/main/java/umc/stockoneqback/auth/domain/model/jwt/AuthToken.java
@@ -1,4 +1,4 @@
-package umc.stockoneqback.auth.domain.model;
+package umc.stockoneqback.auth.domain.model.jwt;
 
 public record AuthToken(
         String accessToken,
