+++ src/main/java/umc/stockoneqback/auth/domain/model/jwt/TokenType.java
@@ -1,0 +1,5 @@
+package umc.stockoneqback.auth.domain.model.jwt;
+
+public enum TokenType {
+    ACCESS, REFRESH
+}
