+++ src/main/java/umc/stockoneqback/auth/domain/repository/FcmTokenRedisRepository.java
@@ -1,7 +1,7 @@
 package umc.stockoneqback.auth.domain.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import umc.stockoneqback.auth.domain.model.FcmToken;
+import umc.stockoneqback.auth.domain.model.fcm.FcmToken;
 
 import java.util.List;
