+++ src/main/java/umc/stockoneqback/auth/domain/repository/TokenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import umc.stockoneqback.auth.domain.model.Token;
+import umc.stockoneqback.auth.domain.model.jwt.Token;
 
 import java.util.Optional;
