+++ src/main/java/umc/stockoneqback/auth/service/jwt/TokenReissueService.java
@@ -3,7 +3,7 @@ package umc.stockoneqback.auth.service.jwt;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import umc.stockoneqback.auth.controller.dto.request.ReissueTokenRequest;
-import umc.stockoneqback.auth.domain.model.AuthToken;
+import umc.stockoneqback.auth.domain.model.jwt.AuthToken;
 import umc.stockoneqback.auth.exception.AuthErrorCode;
 import umc.stockoneqback.auth.utils.TokenProvider;
 import umc.stockoneqback.global.exception.BaseException;
