+++ src/main/java/umc/stockoneqback/product/service/ProductService.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import umc.stockoneqback.auth.domain.model.FcmToken;
+import umc.stockoneqback.auth.domain.model.fcm.FcmToken;
 import umc.stockoneqback.auth.service.fcm.FcmTokenService;
 import umc.stockoneqback.field.domain.store.Store;
 import umc.stockoneqback.field.service.PartTimerService;
